feat: 터지지 않게끔 프래그먼트 분리
